requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.6.1
pymongo==4.10.1
python-jose==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pwdlib[argon2]==0.2.1
python-multipart==0.0.12
email-validator==2.2.0
pytest==9.0.3
pytest-asyncio==1.3.0
httpx==0.28.1